refactor receive permission to LoginFragment
